# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/patients": {
      "post": {
        "description": "Crea el perfil del paciente y vincula al creador como consent-holder. Idempotente por operationId (NFR-34).",
        "operationId": "MembershipController_registerPatient_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPatientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-01 · Registrar paciente",
        "tags": [
          "Membership"
        ]
      },
      "get": {
        "operationId": "MembershipController_myPatients_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-22 · Listar perfiles de paciente de la cuenta",
        "tags": [
          "Membership"
        ]
      }
    },
    "/api/v1/patients/{id}": {
      "get": {
        "description": "Ficha completa. Requiere vínculo con el paciente (cualquier rol); linkRole indica el rol de quien consulta.",
        "operationId": "MembershipController_patientRecord_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientRecordDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-22 · Ver la ficha del paciente",
        "tags": [
          "Membership"
        ]
      },
      "patch": {
        "description": "Set parcial de la ficha (naturalmente idempotente, sin operationId). Solo vínculos consent-holder o manager; queda auditado (quién, cuándo, qué campos).",
        "operationId": "MembershipController_updatePatient_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePatientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientRecordDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-22 · Editar la ficha del paciente",
        "tags": [
          "Membership"
        ]
      }
    },
...
  },
  "info": {
    "title": "Keru API",
    "description": "Backend del MVP de Keru (marketplace de cuidadores). Contrato para clientes (app móvil y web). Ver casos de uso en Keru-Casos-de-Uso-MVP.md y reglas en constitution.md.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "account": {
        "type": "apiKey",
        "in": "header",
        "name": "x-account-id"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "EmergencyContactDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "María Díaz"
          },
          "phone": {
            "type": "string",
            "example": "+54 11 5555-5555"
          },
          "relationship": {
            "type": "string",
            "example": "hija"
          }
        },
        "required": [
          "name",
          "phone"
        ]
      },
      "RegisterPatientDto": {
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string",
            "description": "Identidad de operación provista por el cliente (NFR-34). Un reintento con el mismo valor no duplica el efecto.",
            "example": "op-8f3a2c1e"
          },
          "fullName": {
            "type": "string",
            "example": "Rosa Díaz"
          },
          "birthDate": {
            "type": "string",
            "example": "1948-03-10",
            "description": "Fecha de nacimiento ISO (YYYY-MM-DD). La edad se deriva de acá."
          },
          "photoUrl": {
            "type": "string",
            "example": "https://cdn.keru.app/p/rosa.jpg"
          },
          "mainCondition": {
            "type": "string",
            "example": "Hipertensión"
          },
          "bloodGroup": {
            "type": "string",
            "example": "0+"
          },
          "allergies": {
            "example": [
              "Penicilina"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "emergencyContact": {
            "$ref": "#/components/schemas/EmergencyContactDto"
          }
        },
        "required": [
          "operationId",
          "fullName",
          "birthDate",
          "mainCondition",
          "allergies",
          "emergencyContact"
        ]
      },
      "PatientResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string",
            "example": "Rosa Díaz"
          },
          "age": {
            "type": "number",
            "example": 78,
            "description": "Edad derivada de la fecha de nacimiento."
          },
          "duplicateCandidateId": {
            "type": "string",
            "format": "uuid",
            "description": "Si se detectó un posible duplicado del mismo humano (residuo #21), su id; el cliente puede ofrecer vincular/mergear."
          }
        },
        "required": [
          "id",
          "fullName",
          "age"
        ]
      },
      "PatientRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string",
            "example": "Rosa Díaz"
          },
          "birthDate": {
            "type": "string",
            "example": "1948-03-10",
            "description": "Fecha de nacimiento ISO (YYYY-MM-DD)."
          },
          "age": {
            "type": "number",
            "example": 78,
            "description": "Edad derivada de la fecha de nacimiento."
          },
          "photoUrl": {
            "type": "string",
            "example": "https://cdn.keru.app/p/rosa.jpg"
          },
          "mainCondition": {
            "type": "string",
            "example": "Hipertensión"
          },
          "bloodGroup": {
            "type": "string",
            "example": "0+"
          },
          "allergies": {
            "example": [
              "Penicilina"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "emergencyContact": {
            "type": "object",
            "example": {
              "name": "María Díaz",
              "phone": "+54 11 5555-5555",
              "relationship": "hija"
            }
          },
          "linkRole": {
            "type": "string",
            "enum": [
              "consent-holder",
              "manager",
              "viewer"
            ],
            "description": "Rol del vínculo de la cuenta que consulta."
          }
        },
        "required": [
          "id",
          "fullName",
          "birthDate",
          "age",
          "mainCondition",
          "allergies",
          "emergencyContact",
          "linkRole"
        ]
      },
      "UpdatePatientDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Rosa Díaz"
          },
          "birthDate": {
            "type": "string",
            "example": "1948-03-10",
            "description": "Fecha de nacimiento ISO (YYYY-MM-DD). No puede ser futura."
          },
          "photoUrl": {
            "type": "string",
            "example": "https://cdn.keru.app/p/rosa.jpg"
          },
          "mainCondition": {
            "type": "string",
            "example": "Hipertensión"
          },
          "bloodGroup": {
            "type": "string",
            "example": "0+"
          },
          "allergies": {
            "example": [
              "Penicilina"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "emergencyContact": {
            "$ref": "#/components/schemas/EmergencyContactDto"
          }
        }
      },
...
    }
  }
}